+++ CHANGELOG.md
@@ -2,11 +2,13 @@
 
 ## Not released
 
+## 1.3
+
+### 1.3.0-beta.2 (2022-06-20)
+
 - Add layerOrder in LegendWidget [#433](https://github.com/CartoDB/carto-react/pull/433)
 - Allow FormulaWidget column to be undefined when using COUNT [#434](https://github.com/CartoDB/carto-react/pull/434)
 - Bump deck.gl to 8.8.0-beta.3 [#435](https://github.com/CartoDB/carto-react/pull/435)
-
-## 1.3
 
 ### 1.3.0-beta.1 (2022-06-14)